% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_to_hyper.R
\name{frame_tohyper}
\alias{frame_tohyper}
\alias{frame_to_hyper}
\title{Converting a data.frame to a Tableau hyper extract file}
\usage{
frame_tohyper(d = NULL, f = NULL, tbl = "test")
}
\arguments{
\item{d}{data.frame input}

\item{f}{the filename, with a .hyper extension.  This is required.}
}
\value{
The file is saved
}
\description{
Use this function for to save a data.frame to a Tableau hyper file.
}
\details{
frame_to_hyper


See example below.
}
\references{
\url{https://github.com/jooyoungseo/mboxr/blob/master/R/read_mbox.R}

\url{https://www.anaconda.com/download/}
}
\author{
Tibert, Brock, \email{btibert3@gmail.com}
}
\keyword{frame_to_hyper}
